+++ packages/server/graphql/graphql.ts
@@ -1,8 +1,8 @@
 import {GraphQLFieldConfig} from 'graphql'
 import AuthToken from '../database/types/AuthToken'
+import type {CacheWorker} from '../dataloader/DataLoaderCache'
 import RootDataLoader from '../dataloader/RootDataLoader'
 import {InMemoryRateLimiter} from '../utils/rateLimiters/InMemoryRateLimiter'
-import {CacheWorker} from './DataLoaderCache'
 
 // Avoid needless parsing & validating for the 300 hottest operations
 export type DataLoaderWorker = CacheWorker<RootDataLoader>