+++ packages/server/yoga.ts
@@ -1,0 +1,109 @@
+import {useDeferStream} from '@graphql-yoga/plugin-defer-stream'
+import {usePersistedOperations} from '@graphql-yoga/plugin-persisted-operations'
+import type {GraphQLParams} from 'graphql-yoga'
+import {createYoga, useReadinessCheck} from 'graphql-yoga'
+import {sql} from 'kysely'
+import uws from 'uWebSockets.js'
+import sleep from '../client/utils/sleep'
+import AuthToken from './database/types/AuthToken'
+import getRateLimiter from './graphql/getRateLimiter'
+import {MutationResolvers, QueryResolvers, Resolver} from './graphql/public/resolverTypes'
+import rootSchema from './graphql/public/rootSchema'
+import getKysely from './postgres/getKysely'
+import getVerifiedAuthToken from './utils/getVerifiedAuthToken'
+import {useDatadogTracing} from './utils/useDatadogTracing'
+import {useDisposeDataloader} from './utils/useDisposeDataloader'
+import {usePrivateSchemaForSuperUser} from './utils/usePrivateSchemaForSuperUser'
+
+type OperationResolvers = QueryResolvers & MutationResolvers
+type ExtractArgs<T> = T extends Resolver<any, any, any, infer Args> ? Args : never
+type ExcludedArgs = {
+  [P in keyof OperationResolvers]?: Array<keyof ExtractArgs<OperationResolvers[P]>>
+}
+
+export interface ServerContext {
+  req: uws.HttpRequest
+  res: uws.HttpResponse
+  ip: string
+  authToken: AuthToken | null
+}
+
+export interface UserContext {
+  rateLimiter: ReturnType<typeof getRateLimiter>
+}
+
+export const extractPersistedOperationId = (
+  params: GraphQLParams & {docId?: string; operationName?: string | null}
+) => {
+  return params.docId!
+}
+
+export const getPersistedOperation = async (docId: string) => {
+  if (!__PRODUCTION__) {
+    const queryMap = require('../../queryMap.json')
+    const queryString = queryMap[docId]
+    // for testing we actually commit to the DB & this allows testing on a dev server
+    if (queryString) return queryString
+  }
+  const queryStringRes = await getKysely()
+    .selectFrom('QueryMap')
+    .select('query')
+    .where('id', '=', docId)
+    .executeTakeFirst()
+  return queryStringRes?.query
+}
+
+export const yoga = createYoga<ServerContext, UserContext>({
+  graphqlEndpoint: '/graphql',
+  landingPage: false,
+  plugins: [
+    useDatadogTracing({
+      excludeArgs: {
+        acceptTeamInvitation: ['invitationToken'],
+        loginWithPassword: ['password'],
+        resetPassword: ['token', 'newPassword'],
+        signUpWithPassword: ['password', 'invitationToken'],
+        verifyEmail: ['verificationToken']
+      } as ExcludedArgs,
+      hooks: {
+        execute: (span, args) => {
+          const userId = args.contextValue.authToken?.sub
+          if (userId) {
+            span.setTag('userId', userId)
+          }
+        }
+      },
+      collapse: true
+    }),
+    useDeferStream(),
+    usePersistedOperations({
+      allowArbitraryOperations(request) {
+        const {headers} = request
+        const authHeader = headers.get('authorization')
+        const token = authHeader?.slice(7)
+        const authToken = getVerifiedAuthToken(token)
+        const isSuperUser = authToken?.rol === 'su'
+        return isSuperUser
+      },
+      skipDocumentValidation: true,
+      extractPersistedOperationId,
+      getPersistedOperation
+    }),
+    usePrivateSchemaForSuperUser,
+    useDisposeDataloader,
+    useReadinessCheck({
+      check: async () => {
+        const res = await Promise.race([sql`SELECT 1`.execute(getKysely()), sleep(5000)])
+        if (!res) throw new Error('503')
+      }
+    })
+  ],
+  // There is a bug in graphql-yoga where calling `yoga.getEnveloped` does not work from within graphql-ws when `schema` returns a function
+  // As a workaround, we set the schema via `usePrivateSchemaForSuperUser` using the `onEnveloped` hook
+  schema: rootSchema,
+
+  context: () => {
+    const rateLimiter = getRateLimiter()
+    return {rateLimiter}
+  }
+})