+++ CHANGELOG.md
@@ -5,6 +5,31 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 This CHANGELOG follows conventions [outlined here](http://keepachangelog.com/).
 
+## [9.0.2](https://github.com/ParabolInc/parabol/compare/v9.0.1...v9.0.2) (2025-04-04)
+
+
+### Fixed
+
+* missing githubRequest from schema ([#11105](https://github.com/ParabolInc/parabol/issues/11105)) ([636f4a9](https://github.com/ParabolInc/parabol/commit/636f4a94b480e4103992e296dfb9be5c3700d6bb))
+
+## [9.0.1](https://github.com/ParabolInc/parabol/compare/v9.0.0...v9.0.1) (2025-04-04)
+
+
+### Fixed
+
+* dataloaders in redis bugfix ([#11102](https://github.com/ParabolInc/parabol/issues/11102)) ([7df894d](https://github.com/ParabolInc/parabol/commit/7df894d121e4dc1bf98a19febccae9dad3c2561b))
+
+## [9.0.0](https://github.com/ParabolInc/parabol/compare/v8.39.2...v9.0.0) (2025-04-04)
+
+
+### ⚠ BREAKING CHANGES
+
+* sunset gql-executor service in favor of graphql-yoga on the socket server ([#11077](https://github.com/ParabolInc/parabol/issues/11077))
+
+### Added
+
+* sunset gql-executor service in favor of graphql-yoga on the socket server ([#11077](https://github.com/ParabolInc/parabol/issues/11077)) ([bb592cd](https://github.com/ParabolInc/parabol/commit/bb592cdfce600c04a711352293dfad71f2cefcab))
+
 ## [8.39.2](https://github.com/ParabolInc/parabol/compare/v8.39.1...v8.39.2) (2025-04-04)